# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Healthcare Platform API",
    "description": "Enterprise Healthcare Platform Foundation API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/appointments": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Schedule a new appointment",
        "description": "Book an appointment with conflict detection and resource authorization.",
        "operationId": "create_appointment_api_v1_appointments_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "List appointments with filters and pagination",
        "description": "List appointments. Patients are restricted strictly to their own appointments.",
        "operationId": "list_appointments_api_v1_appointments_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          },
          {
            "name": "provider_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Id"
            }
          },
          {
            "name": "facility_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facility Id"
            }
          },
          {
            "name": "appointment_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Appointment Date"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{appointment_id}": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Get appointment details",
        "description": "Retrieve details for an appointment by UUID (resource-authorized).",
        "operationId": "get_appointment_api_v1_appointments__appointment_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{appointment_id}/reschedule": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Reschedule an appointment",
        "description": "Reschedule an existing scheduled appointment with conflict checks.",
        "operationId": "reschedule_appointment_api_v1_appointments__appointment_id__reschedule_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentRescheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{appointment_id}/cancel": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Cancel an appointment",
        "description": "Cancel an appointment and record cancellation audit.",
        "operationId": "cancel_appointment_api_v1_appointments__appointment_id__cancel_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{appointment_id}/check-in": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Check in patient for today's facility queue",
        "description": "Clinical staff checks in patient, advances appointment, and creates facility QueueEntry.",
        "operationId": "check_in_patient_api_v1_appointments__appointment_id__check_in_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CheckInRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueEntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{appointment_id}/encounter": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Create clinical encounter from appointment",
        "description": "Doctor converts an appointment/queue interaction into an active clinical encounter.",
        "operationId": "create_encounter_from_appointment_api_v1_appointments__appointment_id__encounter_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fhir/Appointment/{appointment_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 Appointment Resource",
        "operationId": "get_fhir_appointment_api_v1_fhir_Appointment__appointment_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Appointment Api V1 Fhir Appointment  Appointment Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppointmentCancelRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "AppointmentCancelRequest"
      },
      "AppointmentCreate": {
        "properties": {
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Facility Id"
          },
          "appointment_date": {
            "type": "string",
            "format": "date",
            "title": "Appointment Date"
          },
          "start_time": {
            "type": "string",
            "format": "time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "time",
            "title": "End Time"
          },
          "appointment_type": {
            "$ref": "#/components/schemas/AppointmentTypeEnum",
            "default": "GENERAL_CONSULTATION"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "provider_id",
          "facility_id",
          "appointment_date",
          "start_time",
          "end_time"
        ],
        "title": "AppointmentCreate"
      },
      "AppointmentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AppointmentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "AppointmentListResponse"
      },
      "AppointmentRescheduleRequest": {
        "properties": {
          "appointment_date": {
            "type": "string",
            "format": "date",
            "title": "Appointment Date"
          },
          "start_time": {
            "type": "string",
            "format": "time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "time",
            "title": "End Time"
          }
        },
        "type": "object",
        "required": [
          "appointment_date",
          "start_time",
          "end_time"
        ],
        "title": "AppointmentRescheduleRequest"
      },
      "AppointmentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Facility Id"
          },
          "appointment_date": {
            "type": "string",
            "format": "date",
            "title": "Appointment Date"
          },
          "start_time": {
            "type": "string",
            "format": "time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "time",
            "title": "End Time"
          },
          "appointment_type": {
            "$ref": "#/components/schemas/AppointmentTypeEnum"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatusEnum"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "cancelled_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled By"
          },
          "cancellation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "provider_id",
          "facility_id",
          "appointment_date",
          "start_time",
          "end_time",
          "appointment_type",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AppointmentResponse"
      },
      "AppointmentStatusEnum": {
        "type": "string",
        "enum": [
          "SCHEDULED",
          "CONFIRMED",
          "CHECKED_IN",
          "WAITING",
          "IN_CONSULTATION",
          "COMPLETED",
          "CANCELLED",
          "NO_SHOW"
        ],
        "title": "AppointmentStatusEnum"
      },
      "AppointmentTypeEnum": {
        "type": "string",
        "enum": [
          "GENERAL_CONSULTATION",
          "FOLLOW_UP",
          "SPECIALIST_REFERRAL",
          "EMERGENCY_TRIAGE",
          "MATERNAL_CHILD_CARE"
        ],
        "title": "AppointmentTypeEnum"
      },
      "CheckInRequest": {
        "properties": {
          "priority": {
            "$ref": "#/components/schemas/QueuePriorityEnum",
            "default": "NORMAL"
          }
        },
        "type": "object",
        "title": "CheckInRequest"
      },
...
      "EncounterResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "appointment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Id"
          },
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Facility Id"
          },
          "encounter_type": {
            "$ref": "#/components/schemas/EncounterTypeEnum"
          },
          "status": {
            "$ref": "#/components/schemas/EncounterStatusEnum"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "chief_complaint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chief Complaint"
          },
          "clinical_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinical Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "provider_id",
          "facility_id",
          "encounter_type",
          "status",
          "started_at",
          "created_at",
          "updated_at"
        ],
        "title": "EncounterResponse"
      },
      "EncounterStatusEnum": {
        "type": "string",
        "enum": [
          "SCHEDULED",
          "IN_PROGRESS",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "EncounterStatusEnum"
      },
      "EncounterTypeEnum": {
        "type": "string",
        "enum": [
          "OUTPATIENT",
          "EMERGENCY",
          "INPATIENT",
          "FOLLOW_UP",
          "HOME_VISIT"
        ],
        "title": "EncounterTypeEnum"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueueEntryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "appointment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Appointment Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Facility Id"
          },
          "queue_date": {
            "type": "string",
            "format": "date",
            "title": "Queue Date"
          },
          "queue_number": {
            "type": "string",
            "title": "Queue Number"
          },
          "priority": {
            "$ref": "#/components/schemas/QueuePriorityEnum"
          },
          "status": {
            "$ref": "#/components/schemas/QueueStatusEnum"
          },
          "checked_in_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked In At"
          },
          "called_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Called At"
          },
          "consultation_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consultation Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "appointment_id",
          "patient_id",
          "facility_id",
          "queue_date",
          "queue_number",
          "priority",
          "status",
          "checked_in_at",
          "created_at"
        ],
        "title": "QueueEntryResponse"
      },
...
      "QueuePriorityEnum": {
        "type": "string",
        "enum": [
          "NORMAL",
          "HIGH",
          "URGENT"
        ],
        "title": "QueuePriorityEnum"
      },
      "QueueStatusEnum": {
        "type": "string",
        "enum": [
          "WAITING",
          "CALLED",
          "IN_CONSULTATION",
          "COMPLETED",
          "SKIPPED",
          "CANCELLED"
        ],
        "title": "QueueStatusEnum"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}